Safari-609.1.11.1/LayoutTests/http/tests/security/contentSecurityPolicy/object-src-url-blocked2.html

<!DOCTYPE html>
<html>
<head>
<script>
if (window.testRunner)
    testRunner.dumpAsText();
</script>
<meta http-equiv="Content-Security-Policy" content="object-src 'none'">
</head>
<body>
<object data="resources/alert-fail.html" type="text/html"></object>
</body>
</html>
