LayoutTests/http/tests/security/contentSecurityPolicy/object-src-url-blocked2-expected.txt

CONSOLE MESSAGE: Refused to load plugin data from 'http://127.0.0.1:8000/security/contentSecurityPolicy/resources/alert-fail.html' because it violates the following Content Security Policy directive: "object-src 'none'".
